[branch-4.1](pick #66515) move orc related serde method to specific datatypes
