+++ be/src/core/data_type_serde/data_type_array_serde.cpp
@@ -444,7 +444,7 @@ Status DataTypeArraySerDe::write_column_to_orc(const std::string& timezone, cons
                                                       packed_nested_size, arena, options));
     // String batches borrow their source bytes, but the packed column is local to this call;
     // keep only those borrowed leaves in the write Arena until Writer::add() consumes them.
-    copy_orc_string_data_to_arena(cur_batch->elements.get(), arena);
+    orc_serde_utils::copy_orc_string_data_to_arena(cur_batch->elements.get(), arena);
     cur_batch->elements->numElements = packed_nested_size;
 
     cur_batch->numElements = end - start;
@@ -665,6 +665,58 @@ bool DataTypeArraySerDe::write_column_to_hive_text(const IColumn& column, Buffer
     }
     bw.write("]", 1);
     return true;
+}
+
+namespace {
+
+Status decode_list_orc_values(const DataTypeSerDeSPtr& nested_serde, IColumn& nested_column,
+                              const OrcDecodedColumnView& orc_view) {
+    const auto* orc_list = dynamic_cast<const ::orc::ListVectorBatch*>(orc_view.batch);
+    if (orc_list == nullptr) {
+        return Status::InternalError("Unexpected ORC list batch type {}",
+                                     orc_view.batch->toString());
+    }
+    DORIS_CHECK(orc_view.file_type != nullptr);
+    DORIS_CHECK(orc_view.selected_type != nullptr);
+    DORIS_CHECK(orc_view.file_type->getSubtypeCount() == 1);
+    DORIS_CHECK(orc_view.selected_type->getSubtypeCount() == 1);
+    DORIS_CHECK(orc_list->elements != nullptr);
+    const auto* file_element_type = orc_view.file_type->getSubtype(0);
+    const auto* selected_element_type = orc_view.selected_type->getSubtype(0);
+    DORIS_CHECK(file_element_type != nullptr);
+    DORIS_CHECK(selected_element_type != nullptr);
+
+    auto& array_column = assert_cast<ColumnArray&>(nested_column);
+    size_t element_size = 0;
+    std::vector<size_t> element_selection;
+    RETURN_IF_ERROR(orc_serde_utils::append_orc_offsets(
+            array_column.get_offsets(), orc_list->offsets, orc_view.rows, &element_size,
+            orc_view.selected_rows, &element_selection));
+    auto element_column = array_column.get_data_ptr()->assert_mutable();
+    const auto child_rows = orc_view.selected_rows == nullptr
+                                    ? element_size
+                                    : static_cast<size_t>(orc_list->elements->numElements);
+    const auto* child_selection = orc_view.selected_rows == nullptr ? nullptr : &element_selection;
+    auto child_view = orc_serde_utils::make_child_orc_view(
+            orc_view, file_element_type, selected_element_type, orc_list->elements.get(),
+            child_rows, child_selection);
+    RETURN_IF_ERROR(
+            orc_serde_utils::read_orc_child_column(nested_serde, element_column, child_view));
+    array_column.get_data_ptr() = std::move(element_column);
+    return Status::OK();
+}
+
+} // namespace
+
+Status DataTypeArraySerDe::read_column_from_orc(IColumn& column,
+                                                const OrcDecodedColumnView& view) const {
+    DORIS_CHECK(view.file_type != nullptr);
+    DORIS_CHECK(view.batch != nullptr);
+    DORIS_CHECK(view.file_type->getKind() == ::orc::TypeKind::LIST);
+    if (orc_serde_utils::orc_decode_row_count(view.rows, view.selected_rows) == 0) {
+        return Status::OK();
+    }
+    return decode_list_orc_values(nested_serde, column, view);
 }
 
 } // namespace doris
