+++ be/src/core/data_type_serde/data_type_decimal_serde.cpp
@@ -37,6 +37,7 @@
 #include "core/data_type/storage_field_type.h"
 #include "core/data_type_serde/arrow_validation.h"
 #include "core/data_type_serde/decoded_column_view.h"
+#include "core/data_type_serde/orc_serde_utils.h"
 #include "core/data_type_serde/parquet_decode_source.h"
 #include "core/types.h"
 #include "exec/common/arithmetic_overflow.h"
@@ -51,6 +52,115 @@
 
 namespace doris {
 namespace {
+
+constexpr int DECIMAL_PRECISION_FOR_HIVE11 = BeConsts::MAX_DECIMAL128_PRECISION;
+
+Int128 to_int128(::orc::Int128 value) {
+    const auto high_bits = static_cast<__uint128_t>(static_cast<uint64_t>(value.getHighBits()));
+    const auto low_bits = static_cast<__uint128_t>(value.getLowBits());
+    return static_cast<Int128>((high_bits << 64) | low_bits);
+}
+
+::orc::Int128 to_orc_int128(Int128 value) {
+    const auto unsigned_value = static_cast<__uint128_t>(value);
+    return ::orc::Int128(static_cast<int64_t>(static_cast<uint64_t>(unsigned_value >> 64)),
+                         static_cast<uint64_t>(unsigned_value));
+}
+
+Status scale_decimal_value(Int128 value, int32_t source_scale, int32_t target_scale,
+                           Int128* scaled_value) {
+    DORIS_CHECK(scaled_value != nullptr);
+    if (source_scale == target_scale) {
+        *scaled_value = value;
+        return Status::OK();
+    }
+    if (source_scale < target_scale) {
+        bool overflow = false;
+        const auto scaled = ::orc::scaleUpInt128ByPowerOfTen(to_orc_int128(value),
+                                                             target_scale - source_scale, overflow);
+        if (overflow) {
+            return Status::DataQualityError(
+                    "ORC decimal value overflows when scaling from {} to {}", source_scale,
+                    target_scale);
+        }
+        *scaled_value = to_int128(scaled);
+        return Status::OK();
+    }
+    *scaled_value = to_int128(
+            ::orc::scaleDownInt128ByPowerOfTen(to_orc_int128(value), source_scale - target_scale));
+    return Status::OK();
+}
+
+void fill_decimal_big_endian_value(Int128 value, std::array<uint8_t, sizeof(Int128)>* bytes) {
+    DORIS_CHECK(bytes != nullptr);
+    const auto unsigned_value = static_cast<__uint128_t>(value);
+    for (size_t byte_idx = 0; byte_idx < bytes->size(); ++byte_idx) {
+        const auto shift = (bytes->size() - byte_idx - 1) * 8;
+        (*bytes)[byte_idx] = static_cast<uint8_t>(unsigned_value >> shift);
+    }
+}
+
+Status decode_decimal_orc_values(const DataTypeSerDe& serde, IColumn& column,
+                                 const OrcDecodedColumnView& orc_view, int32_t target_scale) {
+    DORIS_CHECK(orc_view.file_type != nullptr);
+    auto view = orc_serde_utils::make_orc_decoded_view(orc_view, DecodedValueKind::FIXED_BINARY);
+    view.decimal_precision = orc_view.file_type->getPrecision() == 0
+                                     ? DECIMAL_PRECISION_FOR_HIVE11
+                                     : cast_set<int>(orc_view.file_type->getPrecision());
+    NullMap null_map;
+    orc_serde_utils::fill_orc_decoded_null_map(*orc_view.batch, orc_view.rows,
+                                               orc_view.selected_rows, &null_map);
+    view.null_map = null_map.empty() ? nullptr : null_map.data();
+    view.fixed_length = sizeof(Int128);
+
+    std::vector<StringRef> binary_values;
+    std::vector<std::array<uint8_t, sizeof(Int128)>> decimal_values;
+    const auto output_rows =
+            orc_serde_utils::orc_decode_row_count(orc_view.rows, orc_view.selected_rows);
+    decimal_values.resize(output_rows);
+    binary_values.reserve(output_rows);
+    if (const auto* decimal64_batch =
+                dynamic_cast<const ::orc::Decimal64VectorBatch*>(orc_view.batch);
+        decimal64_batch != nullptr) {
+        view.decimal_scale = decimal64_batch->scale;
+        for (size_t row = 0; row < output_rows; ++row) {
+            Int128 value = 0;
+            const auto source_row = orc_serde_utils::orc_source_row_at(row, orc_view.selected_rows);
+            if (!orc_serde_utils::orc_row_is_null(*orc_view.batch, source_row)) {
+                RETURN_IF_ERROR(scale_decimal_value(decimal64_batch->values[source_row],
+                                                    decimal64_batch->scale, target_scale, &value));
+            }
+            fill_decimal_big_endian_value(value, &decimal_values[row]);
+            binary_values.emplace_back(reinterpret_cast<const char*>(decimal_values[row].data()),
+                                       decimal_values[row].size());
+        }
+        view.binary_values = &binary_values;
+        RETURN_IF_ERROR(orc_serde_utils::read_decoded_values(serde, column, &view));
+        return Status::OK();
+    }
+
+    const auto* decimal128_batch =
+            dynamic_cast<const ::orc::Decimal128VectorBatch*>(orc_view.batch);
+    if (decimal128_batch == nullptr) {
+        return Status::InternalError("Unexpected ORC decimal batch type {}",
+                                     orc_view.batch->toString());
+    }
+    view.decimal_scale = decimal128_batch->scale;
+    for (size_t row = 0; row < output_rows; ++row) {
+        Int128 value = 0;
+        const auto source_row = orc_serde_utils::orc_source_row_at(row, orc_view.selected_rows);
+        if (!orc_serde_utils::orc_row_is_null(*orc_view.batch, source_row)) {
+            RETURN_IF_ERROR(scale_decimal_value(to_int128(decimal128_batch->values[source_row]),
+                                                decimal128_batch->scale, target_scale, &value));
+        }
+        fill_decimal_big_endian_value(value, &decimal_values[row]);
+        binary_values.emplace_back(reinterpret_cast<const char*>(decimal_values[row].data()),
+                                   decimal_values[row].size());
+    }
+    view.binary_values = &binary_values;
+    RETURN_IF_ERROR(orc_serde_utils::read_decoded_values(serde, column, &view));
+    return Status::OK();
+}
 
 template <typename NativeType>
 NativeType decode_big_endian_signed_integer(const uint8_t* data, int length) {
@@ -1336,6 +1446,18 @@ const uint8_t* DataTypeDecimalSerDe<T>::deserialize_binary_to_field(const uint8_
                                "deserialize_binary_to_field with type " + type_to_string(T));
     }
     return data;
+}
+
+template <PrimitiveType T>
+Status DataTypeDecimalSerDe<T>::read_column_from_orc(IColumn& column,
+                                                     const OrcDecodedColumnView& view) const {
+    DORIS_CHECK(view.file_type != nullptr);
+    DORIS_CHECK(view.batch != nullptr);
+    DORIS_CHECK(view.file_type->getKind() == ::orc::TypeKind::DECIMAL);
+    if (orc_serde_utils::orc_decode_row_count(view.rows, view.selected_rows) == 0) {
+        return Status::OK();
+    }
+    return decode_decimal_orc_values(*this, column, view, cast_set<int32_t>(scale));
 }
 
 template class DataTypeDecimalSerDe<TYPE_DECIMAL32>;
