+++ be/src/core/data_type_serde/data_type_struct_serde.cpp
@@ -28,6 +28,7 @@
 #include "core/data_type_serde/arrow_validation.h"
 #include "core/data_type_serde/complex_type_deserialize_util.h"
 #include "core/data_type_serde/data_type_serde.h"
+#include "core/data_type_serde/orc_serde_utils.h"
 #include "core/string_ref.h"
 #include "util/jsonb_document.h"
 #include "util/jsonb_writer.h"
@@ -36,6 +37,58 @@ namespace doris {
 
 class Arena;
 #include "common/compile_check_begin.h"
+
+namespace {
+
+int64_t find_struct_child_index(const ::orc::Type& type, const std::string& field_name) {
+    DORIS_CHECK(type.getKind() == ::orc::TypeKind::STRUCT);
+    for (uint64_t child_idx = 0; child_idx < type.getSubtypeCount(); ++child_idx) {
+        if (type.getFieldName(child_idx) == field_name) {
+            return static_cast<int64_t>(child_idx);
+        }
+    }
+    return -1;
+}
+
+Status decode_struct_orc_values(const DataTypeSerDeSPtrs& elem_serdes_ptrs, IColumn& nested_column,
+                                const OrcDecodedColumnView& orc_view) {
+    const auto* orc_struct = dynamic_cast<const ::orc::StructVectorBatch*>(orc_view.batch);
+    if (orc_struct == nullptr) {
+        return Status::InternalError("Unexpected ORC struct batch type {}",
+                                     orc_view.batch->toString());
+    }
+    DORIS_CHECK(orc_view.file_type != nullptr);
+    DORIS_CHECK(orc_view.selected_type != nullptr);
+    DORIS_CHECK(orc_view.selected_type->getSubtypeCount() == orc_struct->fields.size());
+    auto& struct_column = assert_cast<ColumnStruct&>(nested_column);
+    DORIS_CHECK(struct_column.tuple_size() == orc_view.selected_type->getSubtypeCount());
+    DORIS_CHECK(elem_serdes_ptrs.size() == orc_view.selected_type->getSubtypeCount());
+    for (uint64_t selected_idx = 0; selected_idx < orc_view.selected_type->getSubtypeCount();
+         ++selected_idx) {
+        const auto field_name = orc_view.selected_type->getFieldName(selected_idx);
+        const auto file_child_idx = find_struct_child_index(*orc_view.file_type, field_name);
+        if (file_child_idx < 0) {
+            return Status::InternalError("Selected ORC field {} is not in file struct", field_name);
+        }
+        const auto* file_child_type =
+                orc_view.file_type->getSubtype(static_cast<uint64_t>(file_child_idx));
+        const auto* selected_child_type = orc_view.selected_type->getSubtype(selected_idx);
+        DORIS_CHECK(file_child_type != nullptr);
+        DORIS_CHECK(selected_child_type != nullptr);
+        DORIS_CHECK(selected_idx < orc_struct->fields.size());
+        auto child_column =
+                struct_column.get_column_ptr(static_cast<size_t>(selected_idx))->assert_mutable();
+        auto child_view = orc_serde_utils::make_child_orc_view(
+                orc_view, file_child_type, selected_child_type, orc_struct->fields[selected_idx],
+                orc_view.rows, orc_view.selected_rows);
+        RETURN_IF_ERROR(orc_serde_utils::read_orc_child_column(elem_serdes_ptrs[selected_idx],
+                                                               child_column, child_view));
+        struct_column.get_column_ptr(static_cast<size_t>(selected_idx)) = std::move(child_column);
+    }
+    return Status::OK();
+}
+
+} // namespace
 
 std::string DataTypeStructSerDe::get_name() const {
     size_t size = elem_names.size();
@@ -666,6 +719,17 @@ bool DataTypeStructSerDe::write_column_to_hive_text(const IColumn& column, Buffe
     }
     bw.write("}", 1);
     return true;
+}
+
+Status DataTypeStructSerDe::read_column_from_orc(IColumn& column,
+                                                 const OrcDecodedColumnView& view) const {
+    DORIS_CHECK(view.file_type != nullptr);
+    DORIS_CHECK(view.batch != nullptr);
+    DORIS_CHECK(view.file_type->getKind() == ::orc::TypeKind::STRUCT);
+    if (orc_serde_utils::orc_decode_row_count(view.rows, view.selected_rows) == 0) {
+        return Status::OK();
+    }
+    return decode_struct_orc_values(elem_serdes_ptrs, column, view);
 }
 
 } // namespace doris
